requirements.txt
aiocron==1.8
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
croniter==1.4.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
joblib==1.3.2
libretranslatepy==2.1.1
lxml==4.9.3
magic-filter==1.0.11
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pyaes==1.6.1
pydantic==2.3.0
pydantic_core==2.6.3
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
TgCrypto==1.2.5
tqdm==4.66.1
translate==3.6.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
yarl==1.9.2
